Start work on new previewer
